QualityInspec/requirements.txt
flask
pymssql
pandas
matplotlib
reportlab
python-dotenv
gunicorn
